The Framework is reaching a stabe state.
ModelComponent was tweaked.
In previous commit, values that were to be pre-defined for the View Component, Model Component's Constructor could not be used as GDSource is set after object creation.

This is fixed by implementing a virtual init function for ModelComponent class. This function is called when all references are set in the ModelComponent.