Started with the ID system as many save functionality are dependent on it.
unsigned long long is used to generate ID.
its usecase is decribed in Notes.
With this a more clear Class stucture is created for objects that relate with the offline files and ResourceHandler in perticular.

The Resource classes are only concered with the functions and variables related to the Objects Save and load functionality. The classes can be Expaneded to their fullForm with in their defined Files with additional features and conviniences.